backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
faiss-cpu>=1.7.0
insightface>=0.7.0
onnxruntime>=1.15.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gunicorn>=21.0.0
requests>=2.28.0
pandas>=1.5.0
psutil>=5.9.0
torch>=2.1.0; sys_platform != 'win32'
torchvision>=0.16.0; sys_platform != 'win32'
